undo changes to deposit
