--- src/workflows/depositV4/index.ts
@@ -1,3 +1,0 @@
-export * from './depositEth';
-export * from './depositERC20';
-export * from './depositERC721';
