+++ src/workflows/workflows.ts
@@ -47,10 +47,10 @@ import {
 import { mintingWorkflow } from './minting';
 import { transfersWorkflow, batchTransfersWorkflow } from './transfers';
 import {
-  depositERC20WorkflowV3,
-  depositERC721WorkflowV3,
-  depositEthWorkflowV3,
-} from './depositV3';
+  depositERC20Workflow,
+  depositERC721Workflow,
+  depositEthWorkflow,
+} from './deposit';
 import {
   completeERC20WithdrawalWorkflow,
   completeERC721WithdrawalWorkflow,
@@ -63,11 +63,6 @@ import { generateIMXAuthorisationHeaders } from '../utils';
 import { ImmutableXConfiguration } from '../config';
 import { exchangeTransfersWorkflow } from './exchangeTransfers';
 import axios, { AxiosResponse } from 'axios';
-import {
-  depositEthWorkflowV4,
-  depositERC20WorkflowV4,
-  depositERC721WorkflowV4,
-} from './depositV4';
 
 export class Workflows {
   private readonly depositsApi: DepositsApi;
@@ -195,104 +190,54 @@ export class Workflows {
   }
 
   public async deposit(signer: Signer, deposit: TokenAmount) {
-    const resp = await this.getStarkExContractVersion();
-    const starkExContract = resp.data;
-    const starkExContractVersion = parseInt(starkExContract.version.charAt(0));
     switch (deposit.type) {
       case 'ETH':
-        return this.depositEth(signer, deposit, starkExContractVersion);
+        return this.depositEth(signer, deposit);
       case 'ERC20':
-        return this.depositERC20(signer, deposit, starkExContractVersion);
+        return this.depositERC20(signer, deposit);
       case 'ERC721':
-        return this.depositERC721(signer, deposit, starkExContractVersion);
+        return this.depositERC721(signer, deposit);
     }
   }
 
-  private async depositEth(
-    signer: Signer,
-    deposit: ETHAmount,
-    starkExContractVersion: number,
-  ) {
+  private async depositEth(signer: Signer, deposit: ETHAmount) {
     await this.validateChain(signer);
 
-    if (starkExContractVersion === 3) {
-      return depositEthWorkflowV3(
-        signer,
-        deposit,
-        this.depositsApi,
-        this.usersApi,
-        this.encodingApi,
-        this.config,
-      );
-    } else if (starkExContractVersion >= 4) {
-      return depositEthWorkflowV4(
-        signer,
-        deposit,
-        this.depositsApi,
-        this.encodingApi,
-        this.config,
-      );
-    }
-
-    throw new Error('Unsupported StarkEx version');
+    return depositEthWorkflow(
+      signer,
+      deposit,
+      this.depositsApi,
+      this.usersApi,
+      this.encodingApi,
+      this.config,
+    );
   }
 
-  private async depositERC20(
-    signer: Signer,
-    deposit: ERC20Amount,
-    starkExContractVersion: number,
-  ) {
+  private async depositERC20(signer: Signer, deposit: ERC20Amount) {
     await this.validateChain(signer);
 
-    if (starkExContractVersion === 3) {
-      return depositERC20WorkflowV3(
-        signer,
-        deposit,
-        this.depositsApi,
-        this.usersApi,
-        this.tokensApi,
-        this.encodingApi,
-        this.config,
-      );
-    } else if (starkExContractVersion >= 4) {
-      return depositERC20WorkflowV4(
-        signer,
-        deposit,
-        this.depositsApi,
-        this.tokensApi,
-        this.encodingApi,
-        this.config,
-      );
-    }
-    throw new Error('Unsupported StarkEx version');
+    return depositERC20Workflow(
+      signer,
+      deposit,
+      this.depositsApi,
+      this.usersApi,
+      this.tokensApi,
+      this.encodingApi,
+      this.config,
+    );
   }
 
-  private async depositERC721(
-    signer: Signer,
-    deposit: ERC721Token,
-    starkExContractVersion: number,
-  ) {
+  private async depositERC721(signer: Signer, deposit: ERC721Token) {
     await this.validateChain(signer);
 
-    if (starkExContractVersion === 3) {
-      return depositERC721WorkflowV3(
-        signer,
-        deposit,
-        this.depositsApi,
-        this.usersApi,
-        this.encodingApi,
-        this.config,
-      );
-    } else if (starkExContractVersion >= 4) {
-      return depositERC721WorkflowV4(
-        signer,
-        deposit,
-        this.depositsApi,
-        this.encodingApi,
-        this.config,
-      );
-    }
-    throw new Error('Unsupported StarkEx version');
+    return depositERC721Workflow(
+      signer,
+      deposit,
+      this.depositsApi,
+      this.usersApi,
+      this.encodingApi,
+      this.config,
+    );
   }
 
   public async prepareWithdrawal(
